Module aws_sdk_accessanalyzer::input

Modules

apply_archive_rule_input

See ApplyArchiveRuleInput

cancel_policy_generation_input

See CancelPolicyGenerationInput

create_access_preview_input

See CreateAccessPreviewInput

create_analyzer_input

See CreateAnalyzerInput

create_archive_rule_input

See CreateArchiveRuleInput

delete_analyzer_input

See DeleteAnalyzerInput

delete_archive_rule_input

See DeleteArchiveRuleInput

get_access_preview_input

See GetAccessPreviewInput

get_analyzed_resource_input

See GetAnalyzedResourceInput

get_analyzer_input

See GetAnalyzerInput

get_archive_rule_input

See GetArchiveRuleInput

get_finding_input

See GetFindingInput

get_generated_policy_input

See GetGeneratedPolicyInput

list_access_preview_findings_input

See ListAccessPreviewFindingsInput

list_access_previews_input

See ListAccessPreviewsInput

list_analyzed_resources_input

See ListAnalyzedResourcesInput

list_analyzers_input

See ListAnalyzersInput

list_archive_rules_input

See ListArchiveRulesInput

list_findings_input

See ListFindingsInput

list_policy_generations_input

See ListPolicyGenerationsInput

list_tags_for_resource_input

See ListTagsForResourceInput

start_policy_generation_input

See StartPolicyGenerationInput

start_resource_scan_input

See StartResourceScanInput

tag_resource_input

See TagResourceInput

untag_resource_input

See UntagResourceInput

update_archive_rule_input

See UpdateArchiveRuleInput

update_findings_input

See UpdateFindingsInput

validate_policy_input

See ValidatePolicyInput

Structs

ApplyArchiveRuleInput

Retroactively applies an archive rule.

CancelPolicyGenerationInput

CreateAccessPreviewInput

CreateAnalyzerInput

Creates an analyzer.

CreateArchiveRuleInput

Creates an archive rule.

DeleteAnalyzerInput

Deletes an analyzer.

DeleteArchiveRuleInput

Deletes an archive rule.

GetAccessPreviewInput

GetAnalyzedResourceInput

Retrieves an analyzed resource.

GetAnalyzerInput

Retrieves an analyzer.

GetArchiveRuleInput

Retrieves an archive rule.

GetFindingInput

Retrieves a finding.

GetGeneratedPolicyInput

ListAccessPreviewFindingsInput

ListAccessPreviewsInput

ListAnalyzedResourcesInput

Retrieves a list of resources that have been analyzed.

ListAnalyzersInput

Retrieves a list of analyzers.

ListArchiveRulesInput

Retrieves a list of archive rules created for the specified analyzer.

ListFindingsInput

Retrieves a list of findings generated by the specified analyzer.

ListPolicyGenerationsInput

ListTagsForResourceInput

Retrieves a list of tags applied to the specified resource.

StartPolicyGenerationInput

StartResourceScanInput

Starts a scan of the policies applied to the specified resource.

TagResourceInput

Adds a tag to the specified resource.

UntagResourceInput

Removes a tag from the specified resource.

UpdateArchiveRuleInput

Updates the specified archive rule.

UpdateFindingsInput

Updates findings with the new values provided in the request.

ValidatePolicyInput