Module aws_sdk_accessanalyzer::model

Modules

access_preview

See AccessPreview

access_preview_finding

See AccessPreviewFinding

access_preview_status_reason

See AccessPreviewStatusReason

access_preview_summary

See AccessPreviewSummary

analyzed_resource

See AnalyzedResource

analyzed_resource_summary

See AnalyzedResourceSummary

analyzer_summary

See AnalyzerSummary

archive_rule_summary

See ArchiveRuleSummary

cloud_trail_details

See CloudTrailDetails

cloud_trail_properties

See CloudTrailProperties

criterion

See Criterion

finding

See Finding

finding_source

See FindingSource

finding_source_detail

See FindingSourceDetail

finding_summary

See FindingSummary

generated_policy

See GeneratedPolicy

generated_policy_properties

See GeneratedPolicyProperties

generated_policy_result

See GeneratedPolicyResult

iam_role_configuration

See IamRoleConfiguration

inline_archive_rule

See InlineArchiveRule

internet_configuration

See InternetConfiguration

job_details

See JobDetails

job_error

See JobError

kms_grant_configuration

See KmsGrantConfiguration

kms_grant_constraints

See KmsGrantConstraints

kms_key_configuration

See KmsKeyConfiguration

location

See Location

policy_generation

See PolicyGeneration

policy_generation_details

See PolicyGenerationDetails

position

See Position

s3_access_point_configuration

See S3AccessPointConfiguration

s3_bucket_acl_grant_configuration

See S3BucketAclGrantConfiguration

s3_bucket_configuration

See S3BucketConfiguration

s3_public_access_block_configuration

See S3PublicAccessBlockConfiguration

secrets_manager_secret_configuration

See SecretsManagerSecretConfiguration

sort_criteria

See SortCriteria

span

See Span

sqs_queue_configuration

See SqsQueueConfiguration

status_reason

See StatusReason

substring

See Substring

trail

See Trail

trail_properties

See TrailProperties

validate_policy_finding

See ValidatePolicyFinding

validation_exception_field

See ValidationExceptionField

vpc_configuration

See VpcConfiguration

Structs

AccessPreview

Contains information about an access preview.

AccessPreviewFinding

An access preview finding generated by the access preview.

AccessPreviewStatusReason

Provides more details about the current status of the access preview. For example, if

AccessPreviewSummary

Contains a summary of information about an access preview.

AnalyzedResource

Contains details about the analyzed resource.

AnalyzedResourceSummary

Contains the ARN of the analyzed resource.

AnalyzerSummary

Contains information about the analyzer.

ArchiveRuleSummary

Contains information about an archive rule.

CloudTrailDetails

Contains information about CloudTrail access.

CloudTrailProperties

Contains information about CloudTrail access.

Criterion

The criteria to use in the filter that defines the archive rule.

Finding

Contains information about a finding.

FindingSource

The source of the finding. This indicates how the access that generated the finding is

FindingSourceDetail

Includes details about how the access that generated the finding is granted. This is

FindingSummary

Contains information about a finding.

GeneratedPolicy

Contains the text for the generated policy.

GeneratedPolicyProperties

Contains the generated policy details.

GeneratedPolicyResult

Contains the text for the generated policy and its details.

IamRoleConfiguration

The proposed access control configuration for an IAM role. You can propose a

InlineArchiveRule

An criterion statement in an archive rule. Each archive rule may have multiple

InternetConfiguration

This configuration sets the network origin for the Amazon S3 access point or multi-region

JobDetails

Contains details about the policy generation request.

JobError

Contains the details about the policy generation error.

KmsGrantConfiguration

A proposed grant configuration for a KMS key. For more information, see CreateGrant.

KmsGrantConstraints

Use this structure to propose allowing cryptographic

KmsKeyConfiguration

Proposed access control configuration for a KMS key. You can propose a configuration

Location

A location in a policy that is represented as a path through the JSON representation and

PolicyGeneration

Contains details about the policy generation status and properties.

PolicyGenerationDetails

Contains the ARN details about the IAM entity for which the policy is

Position

A position in a policy.

S3AccessPointConfiguration

The configuration for an Amazon S3 access point or multi-region access point for the bucket.

S3BucketAclGrantConfiguration

A proposed access control list grant configuration for an Amazon S3 bucket. For more

S3BucketConfiguration

Proposed access control configuration for an Amazon S3 bucket. You can propose a

S3PublicAccessBlockConfiguration

The PublicAccessBlock configuration to apply to this Amazon S3 bucket. If the

SecretsManagerSecretConfiguration

The configuration for a Secrets Manager secret. For more information, see CreateSecret.

SortCriteria

The criteria used to sort.

Span

A span in a policy. The span consists of a start position (inclusive) and end position

SqsQueueConfiguration

The proposed access control configuration for an Amazon SQS queue. You can propose a

StatusReason

Provides more details about the current status of the analyzer. For example, if the

Substring

A reference to a substring of a literal string in a JSON document.

Trail

Contains details about the CloudTrail trail being analyzed to generate a policy.

TrailProperties

Contains details about the CloudTrail trail being analyzed to generate a policy.

ValidatePolicyFinding

A finding in a policy. Each finding is an actionable recommendation that can be used to

ValidationExceptionField

Contains information about a validation exception.

VpcConfiguration

The proposed virtual private cloud (VPC) configuration for the Amazon S3 access point. VPC

Enums

AccessPreviewStatus

AccessPreviewStatusReasonCode

AclGrantee

AclPermission

AnalyzerStatus

Configuration

FindingChangeType

FindingSourceType

FindingStatus

FindingStatusUpdate

JobErrorCode

JobStatus

KmsGrantOperation

Locale

NetworkOriginConfiguration

OrderBy

PathElement

PolicyType

ReasonCode

ResourceType

Type

ValidatePolicyFindingType

ValidationExceptionReason