Enum aws_sdk_accessanalyzer::model::Configuration

#[non_exhaustive]
pub enum Configuration {
    IamRole(IamRoleConfiguration),
    KmsKey(KmsKeyConfiguration),
    S3Bucket(S3BucketConfiguration),
    SecretsManagerSecret(SecretsManagerSecretConfiguration),
    SqsQueue(SqsQueueConfiguration),
}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

IamRole(IamRoleConfiguration)

The access control configuration is for an IAM role.

Tuple Fields of IamRole

0: IamRoleConfiguration

KmsKey(KmsKeyConfiguration)

The access control configuration is for a KMS key.

Tuple Fields of KmsKey

0: KmsKeyConfiguration

S3Bucket(S3BucketConfiguration)

The access control configuration is for an Amazon S3 Bucket.

Tuple Fields of S3Bucket

0: S3BucketConfiguration

SecretsManagerSecret(SecretsManagerSecretConfiguration)

The access control configuration is for a Secrets Manager secret.

Tuple Fields of SecretsManagerSecret

0: SecretsManagerSecretConfiguration

SqsQueue(SqsQueueConfiguration)

The access control configuration is for an Amazon SQS queue.

Tuple Fields of SqsQueue

0: SqsQueueConfiguration

Implementations

impl Configuration

pub fn as_iam_role(&self) -> Result<&IamRoleConfiguration, &Self>

pub fn is_iam_role(&self) -> bool

pub fn as_kms_key(&self) -> Result<&KmsKeyConfiguration, &Self>

pub fn is_kms_key(&self) -> bool

pub fn as_s3_bucket(&self) -> Result<&S3BucketConfiguration, &Self>

pub fn is_s3_bucket(&self) -> bool

pub fn as_secrets_manager_secret(
&self
) -> Result<&SecretsManagerSecretConfiguration, &Self>

pub fn is_secrets_manager_secret(&self) -> bool

pub fn as_sqs_queue(&self) -> Result<&SqsQueueConfiguration, &Self>

pub fn is_sqs_queue(&self) -> bool