Module aws_sdk_accessanalyzer::operation

Structs

ApplyArchiveRule

Retroactively applies the archive rule to existing findings that meet the archive rule

CancelPolicyGeneration

Cancels the requested policy generation.

CreateAccessPreview

Creates an access preview that allows you to preview IAM Access Analyzer findings for your

CreateAnalyzer

Creates an analyzer for your account.

CreateArchiveRule

Creates an archive rule for the specified analyzer. Archive rules automatically archive

DeleteAnalyzer

Deletes the specified analyzer. When you delete an analyzer, IAM Access Analyzer is disabled

DeleteArchiveRule

Deletes the specified archive rule.

GetAccessPreview

Retrieves information about an access preview for the specified analyzer.

GetAnalyzedResource

Retrieves information about a resource that was analyzed.

GetAnalyzer

Retrieves information about the specified analyzer.

GetArchiveRule

Retrieves information about an archive rule.

GetFinding

Retrieves information about the specified finding.

GetGeneratedPolicy

Retrieves the policy that was generated using StartPolicyGeneration.

ListAccessPreviewFindings

Retrieves a list of access preview findings generated by the specified access

ListAccessPreviews

Retrieves a list of access previews for the specified analyzer.

ListAnalyzedResources

Retrieves a list of resources of the specified type that have been analyzed by the

ListAnalyzers

Retrieves a list of analyzers.

ListArchiveRules

Retrieves a list of archive rules created for the specified analyzer.

ListFindings

Retrieves a list of findings generated by the specified analyzer.

ListPolicyGenerations

Lists all of the policy generations requested in the last seven days.

ListTagsForResource

Retrieves a list of tags applied to the specified resource.

StartPolicyGeneration

Starts the policy generation request.

StartResourceScan

Immediately starts a scan of the policies applied to the specified resource.

TagResource

Adds a tag to the specified resource.

UntagResource

Removes a tag from the specified resource.

UpdateArchiveRule

Updates the criteria and values for the specified archive rule.

UpdateFindings

Updates the status for the specified findings.

ValidatePolicy

Requests the validation of a policy and returns a list of findings. The findings help